# swagger.yaml
openapi: 3.0.0

info:
  title: Notes API
  description: Notes Management API with Authentication
  version: 1.0.0

servers:
  - url: http://localhost:3000

paths:
  /health:
    get:
      summary: Health Check
      responses:
        "200":
          description: API is healthy

  /auth/signup:
    post:
      summary: Register User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "201":
          description: User created successfully

  /auth/login:
    post:
      summary: Login User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Login successful

  /notes:
    get:
      summary: Get All Notes
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of notes

    post:
      summary: Create Note
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                content:
                  type: string
      responses:
        "201":
          description: Note created successfully

  /notes/{id}:
    get:
      summary: Get Note By ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Note details

    put:
      summary: Update Note
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                content:
                  type: string
      responses:
        "200":
          description: Note updated

    delete:
      summary: Delete Note
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Note deleted

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT